asynccompletion: number of changes when making large changes differ with the original completion
